@@ -1,6 +1,6 @@
 module github.com/netresearch/ldap-manager
 
-go 1.25.1
+go 1.26.0
 
 require (
 	github.com/a-h/templ v0.3.977
