@@ -566,10 +566,8 @@ func TestConcurrentSessionAccess(t *testing.T) {
 	var wg sync.WaitGroup
 	errors := make(chan error, 10)
 
-	for i := 0; i < 10; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
+	for range 10 {
+		wg.Go(func() {
 
 			req := httptest.NewRequest(http.MethodGet, "/protected", http.NoBody)
 			for _, cookie := range cookies {
@@ -586,7 +584,7 @@ func TestConcurrentSessionAccess(t *testing.T) {
 			if resp.StatusCode != http.StatusOK {
 				errors <- fiber.NewError(resp.StatusCode, "unexpected status")
 			}
-		}()
+		})
 	}
 
 	wg.Wait()
